Pushed v7 layout for new release 7.0.0

Following are the features that are going in this release

* Vtiger 7 UI layout

* Easier list views

* Private / Public tags

* Web form attachments

* Variable Taxes, Charges and Regions

* Picklist Coloring

* App Menus

* Rollup comments

* Multiple dashboard tabs

* Controlled sharing of Lists

* Custom Activity Type

* Splitting Edit action in Create / Edit action in roles / profiles.

* Comment with attachments.

See merge request !173

+++ modules/Reports/actions/DeleteAjax.php
@@ -9,6 +9,16 @@
  *************************************************************************************/
 
 class Reports_DeleteAjax_Action extends Vtiger_DeleteAjax_Action {
+    
+	public function checkPermission(Vtiger_Request $request) {
+		$moduleName = $request->getModule();
+		$record = $request->get('record');
+
+		$currentUserPrivilegesModel = Users_Privileges_Model::getCurrentUserPrivilegesModel();
+		if(!$currentUserPrivilegesModel->isPermitted($moduleName, 'Delete', $record)) {
+			throw new AppException(vtranslate('LBL_PERMISSION_DENIED'));
+		}
+	}
 
 	public function process(Vtiger_Request $request) {
 		$moduleName = $request->getModule();
@@ -17,7 +27,7 @@ class Reports_DeleteAjax_Action extends Vtiger_DeleteAjax_Action {
 
 		$recordModel = Reports_Record_Model::getInstanceById($recordId, $moduleName);
 
-		if (!$recordModel->isDefault() && $recordModel->isEditable()) {
+		if (!$recordModel->isDefault() && $recordModel->isEditable() && $recordModel->isEditableBySharing()) {
 			$recordModel->delete();
 			$response->setResult(array(vtranslate('LBL_REPORTS_DELETED_SUCCESSFULLY', $parentModule)));
 		} else {
